+++ app/src/main/java/com/greengarden/ejemplo/prueba.java
@@ -1,6 +1,5 @@
 package com.greengarden.ejemplo;
 
-import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
@@ -12,7 +11,6 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.util.Log;
-import android.view.MotionEvent;
 import android.view.View;
 import android.widget.Button;
 import android.widget.PopupMenu;
@@ -24,9 +22,6 @@ import com.google.firebase.firestore.FirebaseFirestoreException;
 import com.google.firebase.firestore.Query;
 import com.google.firebase.firestore.QuerySnapshot;
 import com.google.gson.Gson;
-import com.greengarden.Consejos.ConsejoAdapter;
-import com.greengarden.Consejos.Consejos;
-import com.greengarden.Consejos.ListaConsejos;
 import com.greengarden.Menu.MenuClickListener;
 import com.greengarden.R;
